add soft delete

+++ app/Http/Controllers/BookController.php
@@ -145,10 +145,12 @@ class BookController extends Controller {
 	 */
 	public function destroy($id)
 	{
-		$this->deleteContent($id);
 		$book = Book::findOrFail($id);
-		$book->contents()->detach();
-		$book->delete();
+		if($book->isOwner() || Auth::user()->isAdmin()) {
+			$this->deleteContent($id);
+			$book->contents()->detach();
+			$book->delete();
+		}
 		return redirect('books');
 	}
 
@@ -185,7 +187,7 @@ class BookController extends Controller {
 		// else
 
 		if($this->alreadyRate($id)){
-			return 'already vote';
+			return redirect('books/'.$id);
 		}
 
 		$userID = Auth::id();
@@ -218,7 +220,7 @@ class BookController extends Controller {
 		$rating -> userKey = $userID;
 		$rating -> save();
 
-		return 'save complete';
+		return redirect('books/'.$id);
 
 	}
